+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/verifier
 
+## 0.12.1
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes